+++ packages/cloud_firestore/cloud_firestore/lib/src/field_value.dart
@@ -14,9 +14,7 @@ class FieldValue extends platform.FieldValuePlatform {
   static final platform.FieldValueFactoryPlatform _factory =
       platform.FieldValueFactoryPlatform.instance;
 
-  FieldValue._(platform.FieldValuePlatform delegate) : super(delegate) {
-    platform.FieldValuePlatform.verifyExtends(delegate);
-  }
+  FieldValue._(dynamic delegate) : super(delegate);
 
   /// Returns a special value that tells the server to union the given elements
   /// with any array value that already exists on the server.
@@ -49,4 +47,15 @@ class FieldValue extends platform.FieldValuePlatform {
   /// server to increment the field’s current value by the given value.
   static FieldValue increment(num value) =>
       FieldValue._(_factory.increment(value));
+
+  @override
+  String toString() => '$runtimeType($delegate)';
+
+  @override
+  bool operator ==(Object o) {
+    return o is FieldValue && o.delegate == delegate;
+  }
+
+  @override
+  int get hashCode => delegate.hashCode;
 }