+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/platform_interface/field_value_factory.dart
@@ -5,6 +5,7 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'package:cloud_firestore_platform_interface/src/method_channel/method_channel_field_value_factory.dart';
+import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 
 /// An interface for a factory that is used to build [FieldValuePlatform] according to
 /// Platform (web or mobile)
@@ -43,7 +44,7 @@ abstract class FieldValueFactoryPlatform extends PlatformInterface {
   /// added to the end. If the field being modified is not already an array it
   /// will be overwritten with an array containing exactly the specified
   /// elements.
-  FieldValuePlatform arrayUnion(List<dynamic> elements) {
+  dynamic arrayUnion(List<dynamic> elements) {
     throw UnimplementedError("arrayUnion() is not implemented");
   }
 
@@ -53,59 +54,24 @@ abstract class FieldValueFactoryPlatform extends PlatformInterface {
   /// All instances of each element specified will be removed from the array.
   /// If the field being modified is not already an array it will be overwritten
   /// with an empty array.
-  FieldValuePlatform arrayRemove(List<dynamic> elements) {
+  dynamic arrayRemove(List<dynamic> elements) {
     throw UnimplementedError("arrayRemove() is not implemented");
   }
 
   /// Returns a sentinel for use with update() to mark a field for deletion.
-  FieldValuePlatform delete() {
+  dynamic delete() {
     throw UnimplementedError("delete() is not implemented");
   }
 
   /// Returns a sentinel for use with set() or update() to include a
   /// server-generated timestamp in the written data.
-  FieldValuePlatform serverTimestamp() {
+  dynamic serverTimestamp() {
     throw UnimplementedError("serverTimestamp() is not implemented");
   }
 
   /// Returns a special value for use with set() or update() that tells the
   /// server to increment the field’s current value by the given value.
-  FieldValuePlatform increment(num value) {
+  dynamic increment(num value) {
     throw UnimplementedError("increment() is not implemented");
   }
-}
-
-/// This is a cross-platform representation of a FieldValue.
-///
-/// Each concrete platform implementation will extend this class,
-/// and add any relevant methods or values to it, so it works as
-/// expected for the target platform.
-///
-/// This is exposed to plugin users, so they can correctly type the
-/// results of calling FieldValueFactory instances, but you should
-/// never "peek" inside subclasses of this.
-///
-/// Just treat it as a "black box".
-class FieldValuePlatform extends PlatformInterface {
-  static final Object _token = Object();
-
-  /// Constructor
-  FieldValuePlatform([this._delegate]) : super(token: _token);
-
-  final FieldValuePlatform _delegate;
-
-  /// Throws an [AssertionError] if [instance] does not extend
-  /// [FieldValuePlatform].
-  ///
-  /// This is used by the app-facing [FieldValue] to ensure that
-  /// the object in which it's going to delegate calls has been
-  /// constructed properly.
-  static verifyExtends(FieldValuePlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
-  }
-
-  /// Used by platform implementers to obtain a value suitable for being passed
-  /// through to the underlying implementation.
-  static dynamic getDelegate(FieldValuePlatform fieldValue) =>
-      fieldValue._delegate;
 }