+++ packages/cloud_firestore/cloud_firestore_platform_interface/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.2]
+
+* Update `FieldValueFactoryPlatform` to produce opaque (`dynamic`) values.
+* `FieldValuePlatform` no longer extends `PlatformInterface`.
+* `MethodChannelFieldValue` now supports equality comparison.
+
 ## [1.0.1]
 
 * Make the pedantic dev_dependency explicit.